chore(rust): split voids_matrix_test into focused suites under 400 lines
Split the 1,228-line voids_matrix_test.rs per the house module-size rule:
- tests/voids_common/{mod,fixtures,production}.rs: shared inline-IFC
  builders and production-path helpers (tests/common subdir pattern,
  not compiled as its own test crate)
- voids_inline_matrix_test.rs: inline_void_matrix
- voids_submesh_test.rs: submesh_void_matrix
- voids_production_test.rs: production fixture tests incl. the Smiley
  sweep canary

All 4 test fns and every assertion preserved; also drops the unused
InlineCtx.uncut field that warned as dead_code.
